La til bekreftelsesdialog før sletting av venner

+++ app/src/main/java/com/example/bursdagsapp/ui/screens/FriendDetailScreen.kt
@@ -4,10 +4,16 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
@@ -22,8 +28,12 @@ fun FriendDetailScreen(
     navController: NavHostController,
     onEditClick: (Friend) -> Unit
 ) {
+    var showDeletePopup by remember { mutableStateOf(false) }
+
     Column(
-        Modifier.fillMaxSize().padding(16.dp),
+        Modifier
+            .fillMaxSize()
+            .padding(16.dp),
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
@@ -39,11 +49,40 @@ fun FriendDetailScreen(
         }
         Button(
             onClick = {
-                viewModel.deleteFriend(friend)
-                navController.navigateUp()
+                showDeletePopup = true
             }
         ) {
             Text("Delete")
         }
+    }
+
+    if (showDeletePopup) {
+        AlertDialog(
+            modifier = Modifier.padding(),
+            onDismissRequest = { showDeletePopup = false},
+            confirmButton = {
+                TextButton(
+                    onClick = {
+                        viewModel.deleteFriend(friend)
+                        navController.navigateUp()
+                    }
+                ) {
+                    Text("Confirm")
+                }
+            },
+            dismissButton = {
+                TextButton(
+                    onClick = { showDeletePopup = false }
+                ) {
+                    Text("Cancel")
+                }
+            },
+            title = {
+                Text("Are you sure?")
+            },
+            text = {
+                Text("Do yo wish to delete ${friend.name}?")
+            }
+        )
     }
 }
